+++ lambda/py/handlers/catcfh_all_exception_handler.py
@@ -1,0 +1,23 @@
+import logging
+from ask_sdk_core.dispatch_components import AbstractExceptionHandler
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+
+
+class CatchAllExceptionHandler(AbstractExceptionHandler):
+    """Catch all exception handler, log exception and
+    respond with custom message.
+    """
+    def can_handle(self, handler_input, exception):
+        # type: (HandlerInput, Exception) -> bool
+        return True
+
+    def handle(self, handler_input, exception):
+        # type: (HandlerInput, Exception) -> Response
+        logger.error(exception, exc_info=True)
+
+        speech = "Sorry, there was some problem. Please try again!!"
+        handler_input.response_builder.speak(speech).ask(speech)
+
+        return handler_input.response_builder.response
