+++ lambda/py/handlers/help_intent_handler.py
@@ -1,0 +1,18 @@
+from ask_sdk_core.dispatch_components import AbstractRequestHandler
+from ask_sdk_core.utils import is_intent_name
+from ask_sdk_model.ui import SimpleCard
+
+class HelpIntentHandler(AbstractRequestHandler):
+    """Handler for Help Intent."""
+    def can_handle(self, handler_input):
+        # type: (HandlerInput) -> bool
+        return is_intent_name("AMAZON.HelpIntent")(handler_input)
+
+    def handle(self, handler_input):
+        # type: (HandlerInput) -> Response
+        speech_text = "You can say hello to me!"
+
+        handler_input.response_builder.speak(speech_text).ask(
+            speech_text).set_card(SimpleCard(
+                "Hello World", speech_text))
+        return handler_input.response_builder.response
