+++ lambda/py/lambda_function.py
@@ -5,128 +5,17 @@
 import logging
 
 from ask_sdk_core.skill_builder import SkillBuilder
-from ask_sdk_core.dispatch_components import AbstractRequestHandler
-from ask_sdk_core.dispatch_components import AbstractExceptionHandler
-from ask_sdk_core.utils import is_request_type, is_intent_name
-from ask_sdk_core.handler_input import HandlerInput
+from handlers.launch_request_handler import LaunchRequestHandler
+from handlers.hello_world_intent_handler import HelloWorldIntentHandler
+from handlers.help_intent_handler import HelpIntentHandler
+from handlers.cancel_or_stop_intent_handler import CancelOrStopIntentHandler
+from handlers.session_ended_request_handler import SessionEndedRequestHandler
+from handlers.catcfh_all_exception_handler import CatchAllExceptionHandler
 
-from ask_sdk_model.ui import SimpleCard
-from ask_sdk_model import Response
-from ask_sdk_model import ui
-from alexa import data
-
-# from ask_sdk_dynamodb.adapter import DynamoDbAdapter
-# dynamodb_adapter = DynamoDbAdapter(
-#     table_name="test_table", partition_keygen=partition_keygen,
-#     dynamodb_resource=dynamodb_resource)
-
-# sb = SkillBuilder(dynamodb_adapter)
 sb = SkillBuilder()
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
-
-
-class LaunchRequestHandler(AbstractRequestHandler):
-    """Handler for Skill Launch."""
-    def can_handle(self, handler_input):
-        # type: (HandlerInput) -> bool
-        return is_request_type("LaunchRequest")(handler_input)
-
-    def handle(self, handler_input):
-        # type: (HandlerInput) -> Response
-        speech_text = data.WELCOME_MESSAGE
-
-        # height = getHeight()
-        # image = getImage(height)
-
-        handler_input.response_builder.speak(speech_text).set_card(
-            ui.StandardCard(
-                title="ことだまさぼてん",
-                text=speech_text,
-                # image=ui.Image(
-                #     small_image_url=small_url,
-                #     large_image_url=large_url
-                # )
-            )
-        )
-        return handler_input.response_builder.response
-
-class HelloWorldIntentHandler(AbstractRequestHandler):
-    """Handler for Hello World Intent."""
-    def can_handle(self, handler_input):
-        # type: (HandlerInput) -> bool
-        return is_intent_name("HelloWorldIntent")(handler_input)
-
-    def handle(self, handler_input):
-        # type: (HandlerInput) -> Response
-        speech_text = "Hello Python World from Classes!"
-
-        handler_input.response_builder.speak(speech_text).set_card(
-            SimpleCard("Hello World", speech_text)).set_should_end_session(
-            True)
-        return handler_input.response_builder.response
-
-
-class HelpIntentHandler(AbstractRequestHandler):
-    """Handler for Help Intent."""
-    def can_handle(self, handler_input):
-        # type: (HandlerInput) -> bool
-        return is_intent_name("AMAZON.HelpIntent")(handler_input)
-
-    def handle(self, handler_input):
-        # type: (HandlerInput) -> Response
-        speech_text = "You can say hello to me!"
-
-        handler_input.response_builder.speak(speech_text).ask(
-            speech_text).set_card(SimpleCard(
-                "Hello World", speech_text))
-        return handler_input.response_builder.response
-
-
-class CancelOrStopIntentHandler(AbstractRequestHandler):
-    """Single handler for Cancel and Stop Intent."""
-    def can_handle(self, handler_input):
-        # type: (HandlerInput) -> bool
-        return (is_intent_name("AMAZON.CancelIntent")(handler_input) or
-                is_intent_name("AMAZON.StopIntent")(handler_input))
-
-    def handle(self, handler_input):
-        # type: (HandlerInput) -> Response
-        speech_text = "Goodbye!"
-
-        handler_input.response_builder.speak(speech_text).set_card(
-            SimpleCard("Hello World", speech_text))
-        return handler_input.response_builder.response
-
-
-class SessionEndedRequestHandler(AbstractRequestHandler):
-    """Handler for Session End."""
-    def can_handle(self, handler_input):
-        # type: (HandlerInput) -> bool
-        return is_request_type("SessionEndedRequest")(handler_input)
-
-    def handle(self, handler_input):
-        # type: (HandlerInput) -> Response
-        return handler_input.response_builder.response
-
-
-class CatchAllExceptionHandler(AbstractExceptionHandler):
-    """Catch all exception handler, log exception and
-    respond with custom message.
-    """
-    def can_handle(self, handler_input, exception):
-        # type: (HandlerInput, Exception) -> bool
-        return True
-
-    def handle(self, handler_input, exception):
-        # type: (HandlerInput, Exception) -> Response
-        logger.error(exception, exc_info=True)
-
-        speech = "Sorry, there was some problem. Please try again!!"
-        handler_input.response_builder.speak(speech).ask(speech)
-
-        return handler_input.response_builder.response
 
 
 sb.add_request_handler(LaunchRequestHandler())
